Formatting + converting a couple more tests to TS
